Replacing integer with bigint.

+++ migrations/2023-04-12-052743_create_history_items/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE history_items (
-  timestamp INTEGER NOT NULL,
+  timestamp BIGINT NOT NULL,
   year INTEGER NOT NULL,
   day INTEGER,
   content TEXT NOT NULL,
